+++ src/Transformers/BM25Transformer.php
@@ -10,7 +10,7 @@ use Rubix\ML\Exceptions\RuntimeException;
 
 use function array_fill;
 use function array_sum;
-use function log;
+use function log1p;
 
 /**
  * BM25 Transformer
@@ -195,7 +195,7 @@ class BM25Transformer implements Transformer, Stateful, Elastic
         $idfs = [];
 
         foreach ($this->dfs as $df) {
-            $idfs[] = log(1.0 + ($this->n - $df + 0.5) / ($df + 0.5));
+            $idfs[] = log1p(($this->n - $df + 0.5) / ($df + 0.5));
         }
 
         $this->idfs = $idfs;